services/ai/requirements.txt
fastapi
uvicorn
python-multipart
requests
Pillow
torch --index-url https://download.pytorch.org/whl/cpu
transformers
qdrant-client
ollama
pydantic
pydantic-settings
sentencepiece 
protobuf